requirements.txt
# Core SC2 Library
burnysc2

protobuf==3.20.3
# Underlying Blizzard Protocol
s2clientprotocol==4.10.1.75800.0

# Networking & Utils
portpicker==1.5.2
websockets==10.4
six==1.16.0

